requirements.txt
pandas
openpyxl
requests
rapidfuzz
rich
python-dotenv
